pyproject.toml
[tool.poetry]
name = "bree"
version = "0.1.0"
description = "Object-oriented keyboard, mouse, and screen control"
authors = ["Michael Lipschultz <michael.lipschultz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyAutoGUI = "^0.9.53"
opencv-python = "^4.5.5"
Pillow = "^9.0.1"
numpy = "^1.22.3"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
